### 2020-06-02 #2196 @kbarnhart
👋 @KMarkert @arbennet @dleutnant @ZhenduoZhu a quick ping to see if you would be interested in reviewing this submission for JOSS.
