### 2020-06-04 #2196 @kbarnhart
👋 @jrkasprzyk @julianneq @nickrsan @dickinsonre, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue which is used to find reviewers. Once 2-3 reviewers have been found I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks.

The submission is about the [pySWMM](https://github.com/OpenWaterAnalytics/pyswmm) package which provides a Python interface to the Stormwater Management Model (SWMM). I'm reaching out to you based on your experience with python, surface water, and/or SWMM. If you are not able to review and have someone to recommend, please mention them here (when mentioning, please place a space after the @ of a github handle, for example, you would refer to me as "@ kbarnhart").

If you are interested in reviewing, I would recommend looking over the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?highlight=conflict%20of%20interest#joss-conflict-of-interest-policy) before the review process starts.

If you have any questions about the JOSS review process, please do not hesitate to reach out to me on this issue or at krbarnhart@usgs.gov
