### 2020-06-04 #2196 @nickrsan
Sure, I'd be happy to - I'd been meaning to volunteer to review for JOSS, but don't think I ever submitted the form, so very timely. My expertise isn't in stormwater - I do work in water systems as katmratliff mentioned though. From the review checklist, it seems like that's not a major issue though?
