### 2020-06-04 #2196 @jrkasprzyk
I’d be happy to, too!

Sent from Mail<https://go.microsoft.com/fwlink/?LinkId=550986> for Windows 10
