### 2020-06-05 #2196 @dickinsonre
I also would be happy to be a reviewer for the PySWMM article.
