### 2020-06-05 #2292 @whedon
[REVIEW]: PySWMM: The Python Interface to Stormwater Management Model (SWMM)
**Submitting author:** @katmratliff (<a href="http://orcid.org/0000-0003-1410-2756">Katherine Ratliff</a>)
**Repository:** <a href="https://github.com/OpenWaterAnalytics/pyswmm" target ="_blank">https://github.com/OpenWaterAnalytics/pyswmm</a>
**Version:** v0.6.2
