+++ paper/paper.bib
@@ -46,7 +46,8 @@
   booktitle={World Environmental and Water Resources Congress 2019: Water, Wastewater, and Stormwater; Urban Water Resources; and Municipal Water Infrastructure},
   pages={46--56},
   year={2019},
-  organization={American Society of Civil Engineers Reston, VA}
+  organization={American Society of Civil Engineers Reston, VA},
+  doi={10.1061/9780784482360.006}
 }
 
 @article{sadler:2019,
@@ -56,7 +57,8 @@
   volume={120},
   pages={104484},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.envsoft.2019.07.009}
 }
 
 @inproceedings{ratliff:2018,
@@ -74,7 +76,8 @@
   number={2},
   pages={04017002},
   year={2017},
-  publisher={American Society of Civil Engineers}
+  publisher={American Society of Civil Engineers},
+  doi={10.1061/JSWBAY.0000817}
 }
 
 @techreport{swmm,
@@ -93,7 +96,7 @@ volume = "113",
 pages = "42 - 47",
 year = "2019",
 issn = "1364-8152",
-doi = "https://doi.org/10.1016/j.envsoft.2018.12.004",
+doi = "10.1016/j.envsoft.2018.12.004",
 url = "http://www.sciencedirect.com/science/article/pii/S1364815218307515",
 author = "Kristina Macro and L. Shawn Matott and Alan Rabideau and Seyed Hamed Ghodsi and Zhenduo Zhu",
 keywords = "Green infrastructure, Multi-objective optimization, SWMM, Rain barrel",
@@ -108,5 +111,6 @@ abstract = "The installation of Green Infrastructure (GI) can revitalize communi
   number={1},
   pages={68--76},
   year={2019},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/1573062X.2019.1611889}
 }
