### 2020-06-22 #2292 @kbarnhart
@katmratliff @nickrsan, @jrkasprzyk, @dickinsonre

I wanted to check in and see how things were going with this review, and to answer any questions if you have them. As mentioned above, JOSS is presently requesting reviewers complete their reviews in six weeks (a bit more than three weeks from now).
