### 2020-06-25 194fc2d
use existing local swmm library before using the embedded one
