### 2020-06-29 #2292 @nickrsan
So, quick note on my status - just ran the test suite and it passed fine on my computer (Python 3.7.7, Win 10). I still have a bit more to do to verify the functional claims, but seems good so far.

One question I have for @kbarnhart - I'm wondering if, in addition to the objective/checklist process for acceptance, if JOSS reviews include general feedback from reviewers. Should we be considering comments of things we think are well done and any suggestions or clarifying questions we have (such as about spots in the documentation), or if they don't affect acceptance, then we should generally not include it here, and any comments should instead be on the target repo? To be clear, I haven't encountered anything significant, but just wanted to confirm - I don't want the authors to miss out on a constructive conversation if I should be attempting to provide that as a reviewer, and alternatively, don't want to be chiming in with anything that's not appropriate here if I should just be following the checklist. It seems like the latter, but I wanted to make sure. Thanks!
