### 2020-06-29 OpenWaterAnalytics/pyswmm#242 @nickrsan
(JOSS Paper) Missing Contributor Information
Hi all,

I'm one of the reviewers for the pyswmm paper with JOSS (https://github.com/openjournals/joss-reviews/issues/2292) - I've almost completed my review and everything seems great, but one thing that seems to be missing that's on the JOSS checklist is contributor guidelines in the documentation - I see the code of conduct in the repository, but JOSS requires the following:

```
There should be clear guidelines for third-parties wishing to:

    Contribute to the software
    Report issues or problems with the software
    Seek support
```
