### 2020-06-29 OpenWaterAnalytics/pyswmm#242 @bemcdonnell
@nickrsan

Thanks for taking time to review! We have put a lot of this info in our wiki - would it make more sense to canonicalize these in the repo itself?

- Contribute to the software

https://github.com/OpenWaterAnalytics/pyswmm/wiki/Author-Contribution-Checklist
https://github.com/OpenWaterAnalytics/pyswmm/wiki/Dev-Github-Workflow

- Report issues or problems with the software / Seek support

Our main README has the following info

Bugs
Our issue tracker is at https://github.com/OpenWaterAnalytics/pyswmm/issues. Please report any bugs that you find. Or, even better, fork the repository on GitHub and create a pull request. All changes are welcome, big or small, and we will help you make the pull request if you are new to git (just ask on the issue).

Can you share an example repository we can model this after? Thanks!
