### 2020-06-29 #2292 @katmratliff
@nickrsan if you have any suggestions/feedback about PySWMM or our documentation beyond the scope of this JOSS review, please feel free to open up issues/PRs in the repo! We are always happy to receive that sort of thing, however big or small, and are always looking to improve! :) thank you for taking the time to review!
