### 2020-06-29 #2292 @kbarnhart
@nickrsan Thanks for a great question. The kind of feedback you are describing is incredibly valuable and I would encourage you to make those comments.

I would recommend that you make issues in the main source repository about these sort of comments and link them to this issue. If you could make a quick note here for me that indicates which issues don't impact acceptance, that would also be very helpful.

Thanks for contributing to the JOSS review process! 👍
