### 2020-06-29 OpenWaterAnalytics/pyswmm#242 @nickrsan
Hi @bemcdonnell - thanks for the details! I'm sorry I missed it in the wiki - I should have thought to check there, but maybe a link from the README would help?

The info in the wiki is sufficient for the paper, as far as I'm concerned. I had been looking for a link in the documentation (which I don't think it needs), in the README, or for a [CONTRIBUTING.md file](https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors) in the repository's root. I personally like the CONTRIBUTING.md files since GitHub prompts people to check it when making pull requests like it does with the code of conduct, but I think it's just one option. For the paper, I think we're set since you have guidelines in a place I think many people would often check, and which you can link people to - I'll go mark this part as complete on the review.
