### 2020-06-30 OpenWaterAnalytics/pyswmm#242 @katmratliff
@nickrsan thank you for the suggestion! I opened up a PR #244 to add the link to the Wiki in the README.
