### 2020-07-13 72792d2
Correcting minor typo in the tutorial document
