### 2020-07-13 #2292 @nickrsan
Review complete - looks great to me and seems like a very valuable tool and contribution - I think the paper clearly states what capabilities the software provides that weren't previously available, and the software performs as expected on my computer.  Overall, it seems like it's in great shape.

@katmratliff I especially liked that the documentation had a tutorial section with multiple sets of code - as a reviewer, it helped me when I got to the phase of verifying the software does what the paper says, but it was also just nice for me, as someone who isn't a stormwater modeler, to see how the package would be used in practice. I'm not sure if it'd be in scope for the package's audience, but it would have been handy if the input files were available for each section (similar to how the `Examples` section approached the input file download) - I pulled ones that seemed appropriate from the project's tests folder though. Even without the files though, it worked well as a demonstration of capabilities. If it seems like something you'd want to include, I can make an issue or a pull request, but since I wasn't sure it'd apply, I thought I'd start by mentioning it here.

Another thing I appreciated is that the code not only had tests, but also references in the README to the current build statuses so I could see it in CI/CD. It really seems like the team is striving for high code quality between that and the submission guidelines.
