### 2020-07-13 #2292 @kbarnhart
Thanks @nickrsan for completing your review, making issues in the repository, and providing a summary here.

@jrkasprzyk, @dickinsonre a friendly reminder to complete your review and/or provide a quick update here on when you anticipate being able to complete it.

Thanks to all for contributing to the JOSS review process.
