### 2020-07-20 #2292 @kbarnhart
@jrkasprzyk if you click the link above and accept the invitation (which expires after some point) then you should be able to edit the checklist. Let me know if it still doesn't address the check/uncheck issue.
