### 2020-07-20 #2292 @jrkasprzyk
Looks like the checkboxes work now, thanks!
