### 2020-07-28 #2292 @jrkasprzyk
My review is complete! I second the comment from @nickrsan about the example files listed. As a person with less GitHub experience, I got a little confused in the Usage statement because it refers to 'model.inp' in the initial example, and I am not sure if it is obvious what that model file is, or the fact that it would need to be in the working directory. It sounds like @nickrsan made an issue about finding the sample files in the documentation, so that is good.
