### 2020-07-29 #2292 @kbarnhart
@jrkasprzyk thanks for completing your review! Glad to hear that changes made in response to @nickrsan 's comments helped with your review.

@dickinsonre could you please provide an update on when you might complete your review.

A note on proceedure: JOSS only requires two reviews, so if I don't hear from @dickinsonre in the next day or two, I will likely remove them as a reviewer and move this submission forward.

Thanks again to all for your participation in this JOSS review.
