### 2020-08-03 #2292 @kbarnhart
@katmratliff not quite yet. I still need to do a final copy-edit on the paper. You should get that from me later today.
