### 2020-08-03 #2292 @kbarnhart
@katmratliff - we are now at the final stage of the JOSS process. PySWMM is a fantastic contribution to surface water modeling and I am excited to see it published in JOSS.

Attached is a PDF with some inline comments. [10.21105.joss.02292_barnhart_edits.pdf](https://github.com/openjournals/joss-reviews/files/5019179/10.21105.joss.02292_barnhart_edits.pdf) Edits/comments pertain to copyediting, reference formatting, acronym definition, and clarity. They are pretty minimal. Once you have addressed them (just link the PR to this thread), then go ahead and do the following steps:

Once those steps are done, ping me here and I'll do my final steps and label this with "recommend accept" which flags the JOSS handling editor in chief who will mange the final publication steps. The EIC may request minor additional changes. Know also that there is a current effort to better standardize author affiliations in JOSS papers (openjournals/joss/issues/783). While nothing has been finalized, the EIC may request that affiliations be expanded (e.g., to include the country). (tagging @openjournals/joss-eics to see if they have any guidance at this point regarding affiliations that the authors of this contribution might address at the same time that they address my copyedits).

As always, let me know if you have any questions/concerns.
