+++ paper/paper.bib
@@ -43,7 +43,7 @@
 @inproceedings{li:2019,
   title={Exploring the Potential for Simulating System-Level Controlled Smart Stormwater System},
   author={Li, Jiada and Burian, Steven and Oroza, Carlos},
-  booktitle={World Environmental and Water Resources Congress 2019: Water, Wastewater, and Stormwater; Urban Water Resources; and Municipal Water Infrastructure},
+  booktitle={World {E}nvironmental and {W}ater {R}esources {C}ongress 2019: {W}ater, {W}astewater, and {S}tormwater; {U}rban {W}ater {R}esources; and {M}unicipal {W}ater {I}nfrastructure},
   pages={46--56},
   year={2019},
   organization={American Society of Civil Engineers Reston, VA},
@@ -90,7 +90,7 @@
 }
 
 @article{macro:2019,
-title = "OSTRICH-SWMM: A new multi-objective optimization tool for green infrastructure planning with SWMM",
+title = "OSTRICH-{S}{W}{M}{M}: {A} {N}ew {M}ulti-{O}bjective {O}ptimization {T}ool for {G}reen {I}nfrastructure {P}lanning with {S}{W}{M}{M}",
 journal = "Environmental Modelling & Software",
 volume = "113",
 pages = "42 - 47",
@@ -104,7 +104,7 @@ abstract = "The installation of Green Infrastructure (GI) can revitalize communi
 }
 
 @article{leutnant:2019,
-  title={swmmr - an R package to interface SWMM},
+  title={swmmr - an {R} package to interface SWMM},
   author={Leutnant, Dominik and D{\"o}ring, Anneke and Uhl, Mathias},
   journal={Urban Water Journal},
   volume={16},
