+++ paper/paper.md
@@ -41,7 +41,7 @@ Stormwater management seeks to reduce runoff from rain or melted snow and improv
 
 The U.S. Environmental Protection Agency's (EPA) [Stormwater Management Model (SWMM)](https://www.epa.gov/water-research/storm-water-management-model-swmm) is a dynamic rainfall-runoff model that has been used for decades to conduct single event or long-term hydrologic, hydraulic, and water quality simulations. It has been used widely throughout the world for planning, analysis, and design of drainage systems [@swmm]. SWMM models the hydrologic processes that generate runoff including rainfall, snowmelt, evaporation, infiltration, and groundwater dynamics, and it routes water through a hydraulic network that includes channels, pipes, storage units, and pumps. The EPA distribution of SWMM does not, however, allow the modeler to interact with the SWMM model during simulation time nor access all of the simulated values and results.
 
-Over the last decade, several libraries have been developed to read, parse, and run SWMM models (\*.inp). These tools have been developed in several programming languages including, but not limited to, Python, R, MATLAB, and Visual Basic. Beyond simple SWMM interfacing utilities, many libraries include a collection of specific features for different applications. Some authors have focused on optimization frameworks [@pathirana:2014; @martinez:2016; @macro:2019] and model calibration [@leutnant:2019], and others have moved toward water quality modeling within their frameworks [@banik:2014]. Even others have extended functionality to support modeling of real-time controls [@riano:2016]. From the most general standpoint, most of these previous works have one major thing in common -- a wrapper interface is used to generate a new \*.inp (SWMM input) file -- which duplicates the data model, adding redundancy. From a maintainability standpoint, the scientific and engineering communities significantly benefit from directly accessing the SWMM data model without having to explicitly read and write new \*.inp files. 
+Over the last decade, several libraries have been developed to read, parse, and run SWMM models (\*.inp). These tools have been developed in several programming languages including, but not limited to, Python, R, MATLAB, and Visual Basic. Beyond simple SWMM interfacing utilities, many libraries include a collection of specific features for different applications. Some authors have focused on optimization frameworks [@pathirana:2014; @martinez:2016; @macro:2019] and model calibration [@leutnant:2019], and others have moved toward water quality modeling within their frameworks [@banik:2014]. Even others have extended functionality to support modeling of real-time controls [@riano:2016]. From the most general standpoint, most of these previous works have one major thing in common -- a wrapper interface is used to generate a new \*.inp (SWMM input) file -- which duplicates the data model, adding redundancy. From a maintainability standpoint, the scientific and engineering communities significantly benefit from directly accessing the SWMM data model during a simulation without having to read and write new \*.inp files. 
 
[…]
