+++ paper/paper.md
@@ -43,7 +43,7 @@ The U.S. Environmental Protection Agency's (EPA) [Stormwater Management Model (S
 
 Over the last decade, several libraries have been developed to read, parse, and run SWMM models (\*.inp). These tools have been developed in several programming languages including, but not limited to, Python, R, MATLAB, and Visual Basic. Beyond simple SWMM interfacing utilities, many libraries include a collection of specific features for different applications. Some authors have focused on optimization frameworks [@pathirana:2014; @martinez:2016; @macro:2019] and model calibration [@leutnant:2019], and others have moved toward water quality modeling within their frameworks [@banik:2014]. Even others have extended functionality to support modeling of real-time controls [@riano:2016]. From the most general standpoint, most of these previous works have one major thing in common -- a wrapper interface is used to generate a new \*.inp (SWMM input) file -- which duplicates the data model, adding redundancy. From a maintainability standpoint, the scientific and engineering communities significantly benefit from directly accessing the SWMM data model during a simulation without having to read and write new \*.inp files. 
 
-In support of the OpenWaterAnalytics  (OWA) open source initiative, the PySWMM project encompasses the development of a Python interfacing wrapper to SWMM with parallel ongoing development of the [Open Water Analytics SWMM5 application programming interface (API)](https://github.com/OpenWaterAnalytics/Stormwater-Management-Model). PySWMM (along with the co-development of the SWMM5-API) is being developed to enable Pythonic to access the SWMM data model, which facilitates rapid prototyping, and also enables users to interact with a model during simulation time. Since PySWMM incorporates enhancements to the SWMM code base, it retains backward compatibilty and ultimately serves to augment what SWMM can do. Current mid-simulation capabilties include the ability to change hydraulic network settings, load externally-generated inflows, and maniuplate low impact development (LID) parameters. Node, link, subcatchment, LID statistics, and results are also accessible both during and after simulation time. PySWMM provides a single framework which encompasses a collection of low-level interfacing functions to the SWMM data model (getters and setters), which facilitates editing of network and hydrologic parameters. This functionality allows researchers and engineers to streamline stormwater model optimization, controls, and result post-processing in order to more effectively address scientific and engineering questions related to water runoff quantity and quality. 
+In support of the OpenWaterAnalytics  (OWA) open source initiative, the PySWMM project encompasses the development of a Python interfacing wrapper to SWMM with parallel ongoing development of the [Open Water Analytics SWMM5 application programming interface (API)](https://github.com/OpenWaterAnalytics/Stormwater-Management-Model). PySWMM (along with the co-development of the SWMM5-API) is being developed to enable Pythonic to access the SWMM data model, which facilitates rapid prototyping, and it also enables users to interact with a model during simulation time. Since PySWMM incorporates enhancements to the SWMM code base, it retains backward compatibilty and ultimately serves to augment what SWMM can do. Current mid-simulation capabilties include the ability to change hydraulic network settings, load externally-generated inflows, and maniuplate low impact development (LID) parameters. Node, link, subcatchment, LID statistics, and results are also accessible both during and after simulation time. PySWMM provides a single framework which encompasses a collection of low-level interfacing functions to the SWMM data model (getters and setters), which facilitates editing of network and hydrologic parameters. This functionality allows researchers and engineers to streamline stormwater model optimization, controls, and result post-processing in order to more effectively address scientific and engineering questions related to water runoff quantity and quality. 
 
[…]
