+++ pyswmm/__init__.py
@@ -17,7 +17,7 @@ from pyswmm.subcatchments import Subcatchment, Subcatchments
 from pyswmm.system import SystemStats
 from pyswmm.raingages import RainGages, RainGage
 
-VERSION_INFO = (0, 6, 1)
+VERSION_INFO = (0, 6, 2, 'dev0')
 __version__ = '.'.join(map(str, VERSION_INFO))
 __swmm_version__ = '5.2.0.dev6'
 __swmm_sha__ = 'bc344c1cf5b5bdf89b3e3435e22c78e229be54a1'
