### 2020-08-04 #2292 @katmratliff
PySWMM [v0.6.1](https://github.com/OpenWaterAnalytics/pyswmm/releases/tag/v0.6.1) has now been released
