### 2020-08-04 #2292 @katmratliff
PySWMM [v0.6.2](https://github.com/OpenWaterAnalytics/pyswmm/releases/tag/v0.6.2) has now been released
