### 2020-08-04 #2292 @kbarnhart
@katmratliff thanks for making those changes and completing these last few steps. I've now marked this as ready for final processing by the handling Editor in Chief.

@nickrsan @jrkasprzyk thank you for your reviews.
